bum ass nigga

+++ src/main/java/today/sleek/client/commands/impl/VClipCommand.java
@@ -13,8 +13,8 @@ public class VClipCommand extends Command {
     @Override
     public void run(String[] args) {
         if (args.length > 0) {
-            ChatUtil.log("Vclipped " + Integer.parseInt(args[0]) + " Blocks");
-            mc.thePlayer.setPosition(mc.thePlayer.posX, mc.thePlayer.posY + Integer.parseInt(args[0]), mc.thePlayer.posZ);
+            ChatUtil.log("Vclipped " + Float.parseFloat(args[0]) + " Blocks");
+            mc.thePlayer.setPosition(mc.thePlayer.posX, mc.thePlayer.posY + Float.parseFloat(args[0]), mc.thePlayer.posZ);
         } else {
             ChatUtil.log(".vclip <amount>");
         }
